backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
pillow>=10.0
numpy>=1.24

# Optional (only if you want to load a saved Keras model):
# tensorflow>=2.12
